refactor: move authSlice to services folder
