+++ src/components/layout/index.tsx
@@ -3,7 +3,7 @@ import { Outlet } from 'react-router-dom';
 import { onAuthStateChanged } from 'firebase/auth';
 import { auth } from '../../firebase.ts';
 import { CircularProgress } from '@mui/material';
-import { authSignIn } from '../../store/auth/authSlice.ts';
+import { authSignIn } from '../../store/services/authSlice.ts';
 import { useAppDispatch } from '../../hooks/hook.ts';
 import Header from '../header';
 import Footer from '../footer';
